feat: 결제수단 드롭다운 바깥 클릭시 닫기 동작
셀렉트 박스와 드롭다운 클릭 시 이벤트 전파를 막고, document 클릭 시 드롭다운 닫는 함수를 호출하도록 함.
blur 이벤트가 정상적으로 동작하지 않아 이 방법으로 대체함.
