+++ README.md
@@ -64,7 +64,7 @@
     - [x] 드롭다운에서 선택지를 클릭하면 셀렉트 박스의 텍스트를 바꾸기
     - [x] 추가하기를 클릭하면 텍스트를 변하지 않도록 처리하기
     - [x] 선택지를 선택하면 셀렉트 박스의 텍스트 색상 변경하기
-    - [ ] 마우스가 드롭다운 패널을 벗어나면 닫기
+    - [x] 드롭다운 패널 바깥을 벗어나면 닫기
     - [ ] x버튼 클릭하면 알림창 띄우기
         - [ ] 클릭시 아이콘을 줄였다 늘려 클릭되는 느낌 주기
         - [ ] 배경 딤 처리
